# openapi/transcendence.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Transcendence API",
    "description": "League analytics API with app API-key and user JWT authentication.",
    "version": "v1"
  },
  "paths": {
...
    "/api/summoners/{summonerId}/stats/overview": {
      "get": {
        "tags": [
          "SummonerStats"
        ],
        "parameters": [
          {
            "name": "summonerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "recent",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerOverviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerOverviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerOverviewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/summoners/{summonerId}/stats/champions": {
      "get": {
        "tags": [
          "SummonerStats"
        ],
        "parameters": [
          {
            "name": "summonerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "top",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionStatDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionStatDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionStatDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/summoners/{summonerId}/stats/roles": {
      "get": {
        "tags": [
          "SummonerStats"
        ],
        "parameters": [
          {
            "name": "summonerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleStatDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleStatDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleStatDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/summoners/{summonerId}/matches/recent": {
      "get": {
        "tags": [
          "SummonerStats"
        ],
        "parameters": [
          {
            "name": "summonerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "queueFamily",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queueIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecentMatchSummaryDtoPagedResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentMatchSummaryDtoPagedResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentMatchSummaryDtoPagedResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/summoners/{summonerId}/matches/{matchId}": {
      "get": {
        "tags": [
          "SummonerStats"
        ],
        "parameters": [
          {
            "name": "summonerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionStatDto": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          },
          "avgKills": {
            "type": "number",
            "format": "double"
          },
          "avgDeaths": {
            "type": "number",
            "format": "double"
          },
          "avgAssists": {
            "type": "number",
            "format": "double"
          },
          "kdaRatio": {
            "type": "number",
            "format": "double"
          },
          "avgCsPerMin": {
            "type": "number",
            "format": "double"
          },
          "avgVisionScore": {
            "type": "number",
            "format": "double"
          },
          "avgDamageToChamps": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "MatchDetailDto": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "nullable": true
          },
          "matchDate": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "queueId": {
            "type": "integer",
            "format": "int32"
          },
          "queueType": {
            "type": "string",
            "nullable": true
          },
          "patch": {
            "type": "string",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantDetailDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MatchRuneDetailDto": {
        "type": "object",
        "properties": {
          "primaryStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "subStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "primarySelections": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "subSelections": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "statShards": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MatchRuneSummaryDto": {
        "type": "object",
        "properties": {
          "primaryStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "subStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "keystoneId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ParticipantDetailDto": {
        "type": "object",
        "properties": {
          "puuid": {
            "type": "string",
            "nullable": true
          },
          "gameName": {
            "type": "string",
            "nullable": true
          },
          "tagLine": {
            "type": "string",
            "nullable": true
          },
          "teamId": {
            "type": "integer",
            "format": "int32"
          },
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "teamPosition": {
            "type": "string",
            "nullable": true
          },
          "win": {
            "type": "boolean"
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "assists": {
            "type": "integer",
            "format": "int32"
          },
          "champLevel": {
            "type": "integer",
            "format": "int32"
          },
          "goldEarned": {
            "type": "integer",
            "format": "int32"
          },
          "totalDamageDealtToChampions": {
            "type": "integer",
            "format": "int32"
          },
          "visionScore": {
            "type": "integer",
            "format": "int32"
          },
          "totalMinionsKilled": {
            "type": "integer",
            "format": "int32"
          },
          "neutralMinionsKilled": {
            "type": "integer",
            "format": "int32"
          },
          "summonerSpell1Id": {
            "type": "integer",
            "format": "int32"
          },
          "summonerSpell2Id": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "runes": {
            "$ref": "#/components/schemas/ParticipantRunesDto"
          }
        },
        "additionalProperties": false
      },
      "ParticipantRunesDto": {
        "type": "object",
        "properties": {
          "primaryStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "subStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "primarySelections": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "subSelections": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "statShards": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RecentMatchSummaryDto": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "nullable": true
          },
          "matchDate": {
            "type": "integer",
            "format": "int64"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "queueId": {
            "type": "integer",
            "format": "int32"
          },
          "queueType": {
            "type": "string",
            "nullable": true
          },
          "win": {
            "type": "boolean"
          },
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "teamPosition": {
            "type": "string",
            "nullable": true
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "assists": {
            "type": "integer",
            "format": "int32"
          },
          "visionScore": {
            "type": "integer",
            "format": "int32"
          },
          "damageToChamps": {
            "type": "integer",
            "format": "int32"
          },
          "csPerMin": {
            "type": "number",
            "format": "double"
          },
          "summonerSpell1Id": {
            "type": "integer",
            "format": "int32"
          },
          "summonerSpell2Id": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "runes": {
            "$ref": "#/components/schemas/MatchRuneSummaryDto"
          },
          "runesDetail": {
            "$ref": "#/components/schemas/MatchRuneDetailDto"
          }
        },
        "additionalProperties": false
      },
      "RecentMatchSummaryDtoPagedResultDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentMatchSummaryDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RecentPerformanceDto": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "nullable": true
          },
          "win": {
            "type": "boolean"
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "assists": {
            "type": "integer",
            "format": "int32"
          },
          "csPerMin": {
            "type": "number",
            "format": "double"
          },
          "visionScore": {
            "type": "integer",
            "format": "int32"
          },
          "damageToChamps": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RoleStatDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "nullable": true
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "SummonerOverviewDto": {
        "type": "object",
        "properties": {
          "summonerId": {
            "type": "string",
            "format": "uuid"
          },
          "totalMatches": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          },
          "avgKills": {
            "type": "number",
            "format": "double"
          },
          "avgDeaths": {
            "type": "number",
            "format": "double"
          },
          "avgAssists": {
            "type": "number",
            "format": "double"
          },
          "kdaRatio": {
            "type": "number",
            "format": "double"
          },
          "avgCsPerMin": {
            "type": "number",
            "format": "double"
          },
          "avgVisionScore": {
            "type": "number",
            "format": "double"
          },
          "avgDamageToChamps": {
            "type": "number",
            "format": "double"
          },
          "avgGameDurationMin": {
            "type": "number",
            "format": "double"
          },
          "recentPerformance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentPerformanceDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "App authentication key for operational endpoints.",
        "name": "X-API-Key",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "User JWT access token from /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}